Clean up of namespaces in Equipment drawing and fixed horizontal lists not showing bottom row

+++ Source/RimWorldHolsters/ModSettingsTools/Selection/Builders/HorizontalBuilder.cs
@@ -48,7 +48,8 @@ namespace ModSettingsTools.Selection.Builders
             _elementWidth = drawRect.width / _elementsPerRow;
 
             int timesButtonCanFit = Mathf.FloorToInt(drawRect.width / _elementWidth);
-            int selectionCountModulo = Mathf.CeilToInt(selection.Count / timesButtonCanFit);
+
+            int selectionCountModulo = Mathf.CeilToInt((float)selection.Count / timesButtonCanFit);
 
             Rect viewRect = new Rect(drawRect.x, drawRect.y, timesButtonCanFit * _elementWidth, buttonHeight * selectionCountModulo);
